requirements.txt
matplotlib==3.5.2
numpy==1.21.4
Pillow==9.2.0
requests==2.27.1
